Create the localstorage file to divide the functions better
